Add CircuitBreakerMultiplexer
It might happen that a particular subsystem only partially fails. For
example, a specific call to an external service might be broken, while
other calls still work. In that case, it might be undesirable to disable
the entire subsystem. This multiplexer allows to dynamically create a
`CircuitBreaker` based on a method argument.